Created a contamination dir. Spread classes to multiple files.
