+++ src/helm/benchmark/contamination/compute_contamination_metrics.py
@@ -2,209 +2,20 @@ import json
 import argparse
 import os
 import glob
-import re
-from dataclasses import dataclass
+
 from typing import List
-from bitarray import bitarray
 from nltk import ngrams
-from typing import Dict, Optional, Generator, DefaultDict
+from typing import Dict, Optional, DefaultDict
 from collections import defaultdict
-from string import punctuation
+
+from helm.benchmark.contamination.light_scenario import LightInstance, LightScenario
+from helm.benchmark.contamination.light_tokenizer import LightTokenizer, DefaultTokenizer
+from helm.benchmark.contamination.document_reading_processor import DocumentReadingProcessor
+from helm.benchmark.contamination.contamination_stats import ContaminationStats, PART_INPUT, PART_REF
+
 
 # The n values of the ngrams to be computed
 N_VALUES: List[int] = [5, 9, 13]  # TODO: Pick the N values
-
-PART_INPUT: str = "input"
-PART_REF: str = "reference"
-
-
-@dataclass(frozen=True, eq=False)
-class LightInstance:
-    """
-    A lighter `Instance` with only text fields.
-    """
-
-    input: str
-    """The input"""
-
-    references: List[str]
-    """References that help us evaluate"""
-
-
-@dataclass(frozen=True, eq=False)
-class LightScenario:
-    """
-    A lighter `Scenario`.
-    """
-
-    scenario_spec: str
-    """The scenario spec"""
-
-    light_instances: List[LightInstance]
-    """Instances of this scenario"""
-
-
-@dataclass(frozen=True, eq=False)
-class DocumentReadingProcessor:
-    """
-    TODO: This is probably a bad class name.
-    """
-
-    file_path: str
-    """The path to the data file"""
-
-    file_format: str
-    """The type of the file format, which determines how it will be read"""
-
-    def get_document_generator(self) -> Generator:
-        if self.file_format == "the_pile":
-            return self.get_the_pile_document_generator()
-        elif self.file_format == "raw":
-            return self.get_raw_document_generator()
-        elif self.file_format == "custom":
-            return self.get_custom_document_generator()
-        else:
-            raise NotImplementedError()
-
-    def get_the_pile_document_generator(self) -> Generator:
-        """
-        This method reads input files with similar file formats with The Pile's jsonl format.
-        Each line of the input file should be a json string, where the document is stored in a field named "text".
-        There are no empty lines between json lines.
-
-        Example:
-        {"text": "Hello World!", "meta": {"pile_set_name": "Pile-CC"}}
-        {"text": "Foo bar", "meta": {"pile_set_name": "Pile-CC"}}
-        """
-        with open(self.file_path, "r") as f:
-            for line in f:
-                yield json.loads(line)["text"]
-
-    def get_raw_document_generator(self) -> Generator:
-        """
-        This method reads input files where each line is a document. The file should not be organized
-        in any specific file structures such as json, jsonl, or tsv, as this may affect ngram computation.
-        Any characters other than the actual text content should be removed.
-
-        Example:
-        Hello World!
-        Foo bar
-        This is the 3rd document.
-        """
-        with open(self.file_path, "r") as f:
-            for line in f:
-                yield line.rstrip("\n")
-
-    def get_custom_document_generator(self) -> Generator:
-        """Define your own document reading method"""
-        pass
-
-
-class LightTokenizer:
-    """
-    Tokenize texts by splitting on whitespaces.
-    """
-
-    def tokenize(self, text: str) -> List[str]:
-        return text.split()
-
-
-class DefaultTokenizer(LightTokenizer):
-    """
-    Normalize and tokenize texts by converting all characters to the lower case and
-    splitting on whitespaces and punctuations.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.r = re.compile(r"[\s{}]+".format(re.escape(punctuation)))
-
-    def tokenize(self, text: str) -> List[str]:
-        return self.r.split(text.lower())
-
-
-class ContaminationStats:
-    """
-    A memory-efficient class for contamination stats. The core data structures are bit arrays where
-    every bit records whether an instance is dirty (contaminated) or not.
-    """
-
-    def __init__(self, scenario_spec: str, num_instances: int, stats_tags: Optional[List[str]] = None):
-        self.scenario_spec = scenario_spec
-        self.num_instances = num_instances
-        self.stats_tags: List[str]
-        if isinstance(stats_tags, list):
-            self.stats_tags = stats_tags
-        else:
-            self.stats_tags = []
-
-        self._input_bits = bitarray(num_instances)
-        self._reference_bits = bitarray(num_instances)
-        self._input_bits.setall(0)
-        self._reference_bits.setall(0)
-
-    @classmethod
-    def from_scenario(cls, scenario: LightScenario, stats_tags: Optional[List[str]] = None):
-        return cls(
-            scenario_spec=scenario.scenario_spec, num_instances=len(scenario.light_instances), stats_tags=stats_tags
-        )
-
-    def write_dirty(self, instance_id: int, part: str):
-        if part == PART_INPUT:
-            self._input_bits[instance_id] = 1
-        elif part == PART_REF:
-            self._reference_bits[instance_id] = 1
-        else:
-            raise ValueError(f"There is no valid part of instance named {part}")
-
-    def get_bit(self, instance_id: int, part: str) -> int:
-        if part == PART_INPUT:
-            return self._input_bits[instance_id]
-        elif part == PART_REF:
-            return self._reference_bits[instance_id]
-        else:
-            raise ValueError(f"There is no valid part of instance named {part}")
-
-    def merge(self, stats):
-        """Merge two stats instance of the same scenario"""
-        if self.scenario_spec != stats.scenario_spec:
-            raise ValueError("Only stats for the same scenario can be merged.")
-        if self.num_instances != stats.num_instances:
-            raise ValueError("The sizes of the two scenarios need to equal.")
-        self._input_bits |= stats._input_bits
-        self._reference_bits |= stats._reference_bits
-
-    @property
-    def num_input_positive_instances(self):
-        return self._input_bits.count()
-
-    @property
-    def num_reference_positive_instances(self):
-        return self._reference_bits.count()
-
-    @property
-    def input_positive_rate(self):
-        return self._input_bits.count() / self.num_instances
-
-    @property
-    def reference_positive_rate(self):
-        return self._reference_bits.count() / self.num_instances
-
-    @property
-    def stats_repr(self) -> str:
-        return f"{self.scenario_spec},{','.join(self.stats_tags)}"
-
-    def generate_summary(self, tags: List[str]) -> str:
-        """Output a summary of the stats"""
-        summary = {
-            "setting": f"{self.stats_repr}{'' if len(tags) == 0 else ','+','.join(tags)}",
-            "total_instances": self.num_instances,
-            "num_input_positive_instances": self.num_input_positive_instances,
-            "num_reference_positive_instances": self.num_reference_positive_instances,
-            "input_positive_rate": self.input_positive_rate,
-            "reference_positive_rate": self.reference_positive_rate,
-        }
-        return json.dumps(summary)
 
 
 def load_scenarios_from_jsonl(path: str) -> List[LightScenario]:
